+++ test/Packman.Test/ManagerTest.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
@@ -58,13 +57,13 @@ namespace Packman.Test
         public async Task InstallPackageWithDefaultPath()
         {
             var entry = await AssemblyLoad.Api.GetInstallablePackage("knockout", "3.1.0");
-            await manager.Install(_manifestPath, entry, Defaults.DefaultLocalPath);
+            await manager.Install(_manifestPath, entry, "lib");
 
             string config = Path.Combine(_cwd, Defaults.ManifestFileName);
             string content = File.ReadAllText(config);
 
             Assert.IsTrue(File.Exists(config), "Config not created");
-            Assert.IsFalse(content.Contains("\"path\": \"" + Defaults.DefaultLocalPath + "\"\r"));
+            Assert.IsFalse(content.Contains("\"path\": \"lib\"\r"));
 
             await manager.Uninstall(config, "knockout");
             content = File.ReadAllText(config);
@@ -75,7 +74,7 @@ namespace Packman.Test
         public async Task InstallPackageDontSaveManifest()
         {
             var entry = await AssemblyLoad.Api.GetInstallablePackage("jquery.ui", "1.11.4");
-            await manager.Install(_manifestPath, entry, Defaults.DefaultLocalPath, false);
+            await manager.Install(_manifestPath, entry, "lib", false);
 
             string config = Path.Combine(_cwd, Defaults.ManifestFileName);